+++ tests/test_log.py
